There was a nasty bug. It turns out that __main__.ClassName is not the same as real_path.ClassName Fixed by extracting common code in the different locations
