+++ backend/dataall/core/catalog/catalog_indexer.py
@@ -1,0 +1,16 @@
+from abc import ABC
+from typing import List
+
+
+class CatalogIndexer(ABC):
+    _INDEXERS: List['CatalogIndexer'] = []
+
+    def __init__(self):
+        CatalogIndexer._INDEXERS.append(self)
+
+    @staticmethod
+    def all():
+        return CatalogIndexer._INDEXERS
+
+    def index(self, session) -> int:
+        raise NotImplementedError("index is not implemented")
