+++ backend/dataall/modules/datasets/tasks/dataset_stack_finder.py
@@ -1,0 +1,19 @@
+import logging
+from typing import List
+
+from dataall.core.stack_finder import StackFinder
+from dataall.modules.datasets_base.db.dataset_repository import DatasetRepository
+from dataall.modules.datasets_base.db.models import Dataset
+
+log = logging.getLogger(__name__)
+
+
+class DatasetStackFinder(StackFinder):
+    """
+    Dataset stack finder. Looks for datasets stack to update
+    Register automatically itself when StackFinder instance is created
+    """
+    def find_stack_uris(self, session) -> List[str]:
+        all_datasets: [Dataset] = DatasetRepository.list_all_active_datasets(session)
+        log.info(f'Found {len(all_datasets)} datasets')
+        return [dataset.datasetUri for dataset in all_datasets]
